+++ src/main.c
@@ -1,227 +1,4 @@
-#include <unistd.h>
-#include <stdlib.h>
-#include <xcb/xcb.h>
-#include <xcb/xcb_keysyms.h>
-#include <xcb/xcb_icccm.h>
-#include <X11/keysym.h>
-
-#include <stdio.h>
-
-#include "action.h"
-#include "xalloc.h"
-#include "util.h"
-
-/* xcb server connection */
-xcb_connection_t    *g_dpy;
-/* screens */
-xcb_screen_t        **g_screens;
-uint32_t            g_screen_count;
-/* the screen being used */
-uint32_t            g_screen_no;
-/* general purpose values */
-uint32_t            g_values[5];
-/* 1 while the window manager is running */
-int g_running;
-
-void init_connection(void)
-{
-    g_dpy = xcb_connect(NULL, NULL);
-    if (xcb_connection_has_error(g_dpy) > 0) {
-        fprintf(stderr, "could not create xcb connection\n");
-        exit(1);
-    }
-}
-
-void log_screens(FILE *fp)
-{
-    xcb_screen_t            *screen;
-
-    fprintf(fp, "Have %u screen(s):\n", g_screen_count);
-    for (uint32_t i = 0; i < g_screen_count; i++) {
-        screen = g_screens[i];
-        fprintf(fp, "Screen %u ; %u:\n", i, screen->root);
-        fprintf(fp, "  width.........: %u\n", screen->width_in_pixels);
-        fprintf(fp, "  height........: %u\n", screen->height_in_pixels);
-        fprintf(fp, "  white pixel...: %u\n", screen->white_pixel);
-        fprintf(fp, "  black pixel...: %u\n", screen->black_pixel);
-        fprintf(fp, "\n");
-    }
-}
-
-void init_screens(void)
-{
-    xcb_screen_iterator_t   iter;
-
-    iter = xcb_setup_roots_iterator(xcb_get_setup(g_dpy));
-    g_screens = xmalloc(sizeof(*g_screens) * iter.rem);
-    g_screen_count = 0;
-    for (; iter.rem > 0; xcb_screen_next(&iter)) {
-        g_screens[g_screen_count++] = iter.data;
-    }
-}
-
-static xcb_keysym_t xcb_get_keysym(xcb_keycode_t keycode)
-{
-    xcb_key_symbols_t *keysyms;
-    xcb_keysym_t       keysym;
-
-    keysyms = xcb_key_symbols_alloc(g_dpy);
-    if (keysyms == NULL) {
-        return 0;
-    }
-    keysym = xcb_key_symbols_get_keysym(keysyms, keycode, 0);
-    xcb_key_symbols_free(keysyms);
-    return keysym;
-}
-
-static xcb_keycode_t *xcb_get_keycodes(xcb_keysym_t keysym)
-{
-    xcb_key_symbols_t   *keysyms;
-    xcb_keycode_t       *keycode;
-
-    keysyms = xcb_key_symbols_alloc(g_dpy);
-    if (keysyms == NULL) {
-        return NULL;
-    }
-    keycode = xcb_key_symbols_get_keycode(keysyms, keysym);
-    xcb_key_symbols_free(keysyms);
-    return keycode;
-}
-
-static struct {
-    int mod;
-    xcb_keysym_t keysym;
-    int action;
-} keys[] = {
-    { XCB_MOD_MASK_1, XK_Return, ACTION_START_TERMINAL },
-};
-
-void grab_keys(void)
-{
-    xcb_screen_t    *screen;
-    xcb_keycode_t   *keycode;
-    uint32_t        modifiers[] = { 0, XCB_MOD_MASK_LOCK };
-
-    screen = g_screens[0];
-    xcb_ungrab_key(g_dpy, XCB_GRAB_ANY, screen->root, XCB_MOD_MASK_ANY);
-    for (uint32_t i = 0; i < SIZE(keys); i++) {
-        keycode = xcb_get_keycodes(keys[i].keysym);
-        for (uint32_t k = 0; keycode[k] != XCB_NO_SYMBOL; k++) {
-            for (uint32_t m = 0; m < SIZE(modifiers); m++) {
-                xcb_grab_key(g_dpy, 1, screen->root, keys[i].mod | modifiers[m],
-                        keycode[k], XCB_GRAB_MODE_ASYNC, XCB_GRAB_MODE_ASYNC);
-            }
-        }
-        free(keycode);
-    }
-}
-
-int take_control(int screen_no)
-{
-    xcb_screen_t        *screen;
-    xcb_generic_error_t *error;
-
-    screen = g_screens[screen_no];
-    g_screen_no = screen_no;
-    g_values[0] = XCB_EVENT_MASK_SUBSTRUCTURE_REDIRECT |
-        XCB_EVENT_MASK_STRUCTURE_NOTIFY |
-        XCB_EVENT_MASK_SUBSTRUCTURE_NOTIFY |
-        XCB_EVENT_MASK_BUTTON_PRESS |
-        XCB_EVENT_MASK_PROPERTY_CHANGE;
-    error = xcb_request_check(g_dpy, xcb_change_window_attributes_checked(g_dpy, screen->root,
-            XCB_CW_EVENT_MASK, g_values));
-    if (error != NULL) {
-        return 1;
-    }
-    //xcb_ungrab_key(g_dpy, XCB_GRAB_ANY, screen->root, XCB_MOD_MASK_ANY);
-    /*int key_table_size = sizeof(keys) / sizeof(*keys);
-    for (int i = 0; i < key_table_size; ++i) {
-        xcb_keycode_t *keycode = xcb_get_keycodes(keys[i].keysym);
-        if (keycode != NULL) {
-            xcb_grab_key(dpy, 1, scre->root, keys[i].mod, *keycode,
-                XCB_GRAB_MODE_ASYNC, XCB_GRAB_MODE_ASYNC );
-        }
-    }*/
-    xcb_flush(g_dpy);
-    /*xcb_grab_button(g_dpy, 0, screen->root, XCB_EVENT_MASK_BUTTON_PRESS |
-        XCB_EVENT_MASK_BUTTON_RELEASE, XCB_GRAB_MODE_ASYNC,
-        XCB_GRAB_MODE_ASYNC, screen->root, XCB_NONE, 1, MOD1);
-    xcb_grab_button(g_dpy, 0, screen->root, XCB_EVENT_MASK_BUTTON_PRESS |
-        XCB_EVENT_MASK_BUTTON_RELEASE, XCB_GRAB_MODE_ASYNC,
-        XCB_GRAB_MODE_ASYNC, screen->root, XCB_NONE, 3, MOD1);*/
-    xcb_flush(g_dpy);
-    return 0;
-}
-
-void set_focus_window(xcb_drawable_t win)
-{
-    xcb_set_input_focus(g_dpy, XCB_INPUT_FOCUS_POINTER_ROOT, win,
-            XCB_CURRENT_TIME);
-}
-
-void accept_new_window(xcb_drawable_t win, xcb_screen_t *screen)
-{
-    xcb_map_window(g_dpy, win);
-    g_values[0] = 0;
-    g_values[1] = 0;
-    g_values[2] = screen->width_in_pixels;
-    g_values[3] = screen->height_in_pixels;
-    g_values[4] = 0;
-    xcb_configure_window(g_dpy, win, XCB_CONFIG_WINDOW_X |
-        XCB_CONFIG_WINDOW_Y | XCB_CONFIG_WINDOW_WIDTH |
-        XCB_CONFIG_WINDOW_HEIGHT | XCB_CONFIG_WINDOW_BORDER_WIDTH, g_values);
-    xcb_flush(g_dpy);
-    g_values[0] = XCB_EVENT_MASK_ENTER_WINDOW | XCB_EVENT_MASK_FOCUS_CHANGE;
-    xcb_change_window_attributes_checked(g_dpy, win,
-        XCB_CW_EVENT_MASK, g_values);
-    set_focus_window(win);
-}
-
-void handle_key_press(xcb_key_press_event_t *ev)
-{
-    xcb_keysym_t           keysym;
-
-    keysym = xcb_get_keysym(ev->detail);
-    for (uint32_t i = 0; i < SIZE(keys); i++) {
-        if (keysym == keys[i].keysym && keys[i].mod == ev->state) {
-            fprintf(stderr, "triggered keybind: %d\n", keys[i].action);
-            do_action(keys[i].action);
-        }
-    }
-}
-
-void log_event(xcb_generic_event_t *ev, FILE *fp);
-
-void handle_event(void)
-{
-    xcb_generic_event_t *ev;
-
-    if (xcb_connection_has_error(g_dpy) > 0) {
-        /* quit because the connection is broken */
-        g_running = 0;
-        return;
-    }
-    ev = xcb_wait_for_event(g_dpy);
-    if (ev != NULL) {
-        log_event(ev, stderr);
-        fputc('\n', stderr);
-        switch ((ev->response_type & ~0x80)) {
-        case XCB_MAP_REQUEST:
-            accept_new_window(((xcb_map_request_event_t*) ev)->window, g_screens[0]);
-            break;
-
-        case XCB_KEY_PRESS:
-            handle_key_press((xcb_key_press_event_t*) ev);
-            break;
-        }
-        free(ev);
-    }
-}
-
-void close_connection(void)
-{
-    xcb_disconnect(g_dpy);
-}
+#include "fensterchef.h"
 
 int main(void)
 {
@@ -229,7 +6,9 @@ int main(void)
     init_screens();
     grab_keys();
     if (g_screen_count > 0) {
-        g_running = !take_control(0);
+        /* take the first screen */
+        g_screen_no = 0;
+        g_running = !take_control();
     }
 
     while (g_running) {
